refactor:지난주 로또생성도 lotto가 책임진다.
