+++ src/main/java/nextstep/step2/domain/Lotto.java
@@ -1,12 +1,18 @@
 package nextstep.step2.domain;
 
+import java.util.Collections;
 import java.util.Set;
+import java.util.stream.Collectors;
+import java.util.stream.Stream;
 
 public class Lotto {
 	private static final String WRONG_LOTTO_NUMBERS = "로또 번호는 6개여야 합니다.";
+	private static final String WRONG_WINNER_NUMBERS = "지난주 로또 당첨번호는 모두 숫자여야 합니다.";
 	protected static final int LOTTO_SIZE = 6;
 	private Set<LottoNumber> numbers;
 
+	public Lotto() {
+	}
 	public Lotto(Set<LottoNumber> numbers) {
 		this.numbers = numbers;
 	}
@@ -18,6 +24,23 @@ public class Lotto {
 	public void validate() {
 		if (numbers.size() != LOTTO_SIZE) {
 			throw new IllegalArgumentException(WRONG_LOTTO_NUMBERS);
+		}
+	}
+
+	public Lotto getLastWeekLotto(String winnerNumbers) {
+		String[] winningNumberSplit = winnerNumbers.replaceAll(" ", "").split(",");
+		validateLastWinnerNumbers(winningNumberSplit);
+		Set<LottoNumber> winningNumbers =  Stream.of(winningNumberSplit)
+				.map(Integer::parseInt)
+				.map(LottoNumber::new)
+				.collect(Collectors.collectingAndThen(Collectors.toSet(), Collections::unmodifiableSet));
+		return new Lotto(winningNumbers);
+	}
+
+	protected void validateLastWinnerNumbers(String[] numbers) {
+		boolean isDigit = Stream.of(numbers).allMatch(oddNum -> oddNum.chars().allMatch(Character::isDigit));
+		if (!isDigit) {
+			throw new IllegalArgumentException(WRONG_WINNER_NUMBERS);
 		}
 	}
 }
